refactor(ui): move specialCharacters const to utils.ts inside password-input folder
