+++ src/scalars/lib/utils.ts
@@ -9,5 +9,3 @@ export function cn(...inputs: ClassValue[]) {
 export const useUniqueId = () => {
   return useId()
 }
-
-export const specialCharacters = ' !"#$%&\'()*+,-./:;<?>@=[\\]^_`{|}~'
