+++ src/ui/components/data-entry/password-input/utils.ts
@@ -1,0 +1,1 @@
+export const specialCharacters = ' !"#$%&\'()*+,-./:;<?>@=[\\]^_`{|}~'
